+++ internal/services/sockets/server.go
@@ -1,10 +1,18 @@
 package sockets
 
-import "net/http"
+import (
+	"net/http"
+	"time"
 
-func NewServer() *http.Server {
-	return &http.Server{
-		Addr:    ":80", // TODO: set addr with a configurable var.
-		Handler: nil,   // TODO: set handler with socket router.
-	}
+	"github.com/ciph-r/postage/internal/services"
+)
+
+func NewService() services.Service {
+	return services.NewHTTP(
+		&http.Server{
+			Addr:    ":80", // TODO: set addr with a configurable var.
+			Handler: nil,   // TODO: set handler with socket router.
+		},
+		time.Minute,
+	)
 }
